+++ Components/WAColorPicker.razor.cs
@@ -68,6 +68,12 @@ namespace WebAwesomeBlazor.Components
         [Parameter]
         public string? Swatches { get; set; }
 
+        /// <summary>
+        /// The preferred placement of the color picker's popup. Note that the actual placement will vary as configured to keep the panel inside of the viewport.
+        /// </summary>
+        [Parameter]
+        public PickerPlacement Placement { get; set; } = PickerPlacement.BottomStart;
+
         #endregion
 
         #region Computed Properties
@@ -97,6 +103,29 @@ namespace WebAwesomeBlazor.Components
                     PickerSize.Large => "large",
                     PickerSize.Inherit => "inherit",
                     _ => "inherit"
+                };
+            }
+        }
+
+        string PlacementString
+        {
+            get
+            {
+                return Placement switch
+                {
+                    PickerPlacement.TopStart => "top-start",
+                    PickerPlacement.Top => "top",
+                    PickerPlacement.TopEnd => "top-end",
+                    PickerPlacement.RightStart => "right-start",
+                    PickerPlacement.Right => "right",
+                    PickerPlacement.RightEnd => "right-end",
+                    PickerPlacement.BottomStart => "bottom-start",
+                    PickerPlacement.Bottom => "bottom",
+                    PickerPlacement.BottomEnd => "bottom-end",
+                    PickerPlacement.LeftStart => "left-start",
+                    PickerPlacement.Left => "left",
+                    PickerPlacement.LeftEnd => "left-end",
+                    _ => "bottom-start"
                 };
             }
         }