+++ docs/CHANGELOG.md
@@ -1,5 +1,18 @@
 ﻿# Changelog
 
+## Version 1.5.0
+- Upgraded to [Web Awesome 3.5.0](https://webawesome.com/docs/resources/changelog#wa_350) - Make sure you have updated your project version on [Web Awesome Teams](https://webawesome.com/teams).
+- Breaking: Project initialisation has changed. Refer to the Readme for details.
+- Removed Pro from [WAPage](/docs/WAPage.md) and [WALayoutContent](/docs/WALayoutContent.md) as they are now available to all users.
+- Added new component: [WAMarkdown](/docs/WAMarkdown.md) - The markdown component turns raw markdown into rendered HTML using the Marked library.
+- Addded new parameter to [WAColorPicker](/docs/WAColorPicker.md): `Placement` to define the preferred placement of the color picker's popup.
+- Added default slot to [WAColorPicker](/docs/WAColorPicker.md) to allow for custom trigger content.
+- Added new parameters to [WATextArea](/docs/WATextArea.md):
+   * `ShowCount` to display a character count below the textarea.
+   * `MaxLength` to specify the maximum number of characters allowed in the textarea.
+   * `MinLength` to specify the minimum number of characters required in the textarea.
+
+
 ## Version 1.4.0
 - Upgraded to [Web Awesome 3.4.0](https://webawesome.com/docs/resources/changelog#wa_340) - Make sure you have updated your project version on [Web Awesome Teams](https://webawesome.com/teams).
 - [WAColorPicker](/docs/WAColorPicker.md): Added new method `SetSwatchesAsync` to set color swatches with accessible labels to display as presets.